Rename from mxgconv to mxpeg-decoder to not have dll conflicts with old player
