+++ mxpeg-decoder/mxpeg-decoder.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mxgconv", "mxgconv.vcxproj", "{5034E99D-C243-3693-963F-03EBE053EAB6}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mxpeg-decoder", "mxpeg-decoder.vcxproj", "{5034E99D-C243-3693-963F-03EBE053EAB6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
