+++ mxpeg-decoder/version.pri
